:rocket: Producto Final Version 1.0 :rocket:
Modificacion de vectores para los tipos de columnas, Lanzamiento del .jar para ejecución.

+++ src/app/pvgps/inventario/ModuloProdBajosDeInv.java
@@ -187,8 +187,8 @@ public void prepararVista( String modulo )
             
             vecTiposColumnas.add     ( EjecutorSQL.STRING);
             vecTiposColumnas.add     ( EjecutorSQL.STRING);
-            vecTiposColumnas.add     ( EjecutorSQL.DOUBLE);
-            vecTiposColumnas.add     ( EjecutorSQL.DOUBLE);
+            vecTiposColumnas.add     ( EjecutorSQL.FLOAT);
+            vecTiposColumnas.add     ( EjecutorSQL.FLOAT);
             vecTiposColumnas.add     ( EjecutorSQL.INT);
             
         }
